Add a generic version of tunnel settings
