+++ ios/MullvadVPN/SettingsManager/TunnelSettings.swift
@@ -1,0 +1,12 @@
+//
+//  TunnelSettings.swift
+//  MullvadVPN
+//
+//  Created by Marco Nikic on 2023-07-31.
+//  Copyright © 2023 Mullvad VPN AB. All rights reserved.
+//
+
+import Foundation
+
+/// Alias to the latest version of the `TunnelSettings`.
+typealias LatestTunnelSettings = TunnelSettingsV2
