+++ ios/MullvadVPN/SettingsManager/TunnelSettingsV2.swift
@@ -33,115 +33,6 @@ struct TunnelSettingsV2: Codable, Equatable {
     var dnsSettings = DNSSettings()
 }
 
-struct StoredAccountData: Codable, Equatable {
-    /// Account identifier.
-    var identifier: String
-
-    /// Account number.
-    var number: String
-
-    /// Account expiry.
-    var expiry: Date
-
-    /// Set to `true` when the account is created and flipped to `false` when the user adds more credit.
-    var isNew = false
-
-    /// Returns `true` if account has expired.
-    var isExpired: Bool {
-        expiry <= Date()
-    }
-}
-
-extension StoredAccountData {
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.identifier = try container.decode(String.self, forKey: .identifier)
-        self.number = try container.decode(String.self, forKey: .number)
-        self.expiry = try container.decode(Date.self, forKey: .expiry)
-
-        // When the app is upgraded from 2023.3 or below, this field won't exist, and the auto synthesized init will fail.
-        // This leads to a reset of the settings. If the key isn't present, consider the account not new to avoid the issue.
-        let isNewAccount = try? container.decode(Bool.self, forKey: .isNew)
-        self.isNew = isNewAccount ?? false
-    }
-}
-
-enum DeviceState: Codable, Equatable {
-    case loggedIn(StoredAccountData, StoredDeviceData)
-    case loggedOut
-    case revoked
-
-    private enum LoggedInCodableKeys: String, CodingKey {
-        case _0 = "account"
-        case _1 = "device"
-    }
-
-    var isLoggedIn: Bool {
-        switch self {
-        case .loggedIn:
-            return true
-        case .loggedOut, .revoked:
-            return false
-        }
-    }
-
-    var accountData: StoredAccountData? {
-        switch self {
-        case let .loggedIn(accountData, _):
-            return accountData
-        case .loggedOut, .revoked:
-            return nil
-        }
-    }
-
-    var deviceData: StoredDeviceData? {
-        switch self {
-        case let .loggedIn(_, deviceData):
-            return deviceData
-        case .loggedOut, .revoked:
-            return nil
-        }
-    }
-}
-
-struct StoredDeviceData: Codable, Equatable {
-    /// Device creation date.
-    var creationDate: Date
-
-    /// Device identifier.
-    var identifier: String
-
-    /// Device name.
-    var name: String
-
-    /// Whether relay hijacks DNS from this device.
-    var hijackDNS: Bool
-
-    /// IPv4 address + mask assigned to device.
-    var ipv4Address: IPAddressRange
-
-    /// IPv6 address + mask assigned to device.
-    var ipv6Address: IPAddressRange
-
-    /// WireGuard key data.
-    var wgKeyData: StoredWgKeyData
-
-    /// Returns capitalized device name.
-    var capitalizedName: String {
-        name.capitalized
-    }
-
-    /// Fill in part of the structure that contains device related properties from `Device` struct.
-    mutating func update(from device: Device) {
-        identifier = device.id
-        name = device.name
-        creationDate = device.created
-        hijackDNS = device.hijackDNS
-        ipv4Address = device.ipv4Address
-        ipv6Address = device.ipv6Address
-    }
-}
-
 struct StoredWgKeyData: Codable, Equatable {
     /// Private key creation date.
     var creationDate: Date
